apps/clinicdesk/requirements.txt
Flask==3.0.3
Faker==24.8.0
python-dotenv==1.0.1
